pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "arabic-reshaper>=3.0.0",
    "librosa>=0.11.0",
    "moviepy>=2.2.1",
    "numpy>=2.3.3",
    "pillow>=11.3.0",
    "pysrt>=1.1.2",
    "python-bidi>=0.6.6",
    "streamlit>=1.50.0",
]
